tests/requirements.txt
# Test requirements for Wildfire Watch
# Python 3.12 compatible versions

# Testing framework
pytest>=8.0.0
pytest-asyncio>=0.21.0
pytest-timeout>=2.1.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0

# MQTT testing
paho-mqtt>=2.0.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Network testing
requests>=2.31.0
aiohttp>=3.9.0

# Data processing
numpy>=1.24.0
opencv-python>=4.8.0

# Certificate testing
cryptography>=41.0.0

# GPIO mocking (for non-Pi systems)
fake-rpi>=0.7.1

# Camera discovery testing
onvif-zeep>=0.2.12
wsdiscovery>=2.1.2
zeep>=4.2.0
netifaces>=0.11.0
scapy>=2.5.0

# Utility
# Note: ipaddress, json, and unittest.mock are built-in Python modules